Refactoring `packages/wallet`
