+++ packages/wallet/src/test-utils.ts
@@ -1,0 +1,2 @@
+export * from './test-utils/generateTestWallet';
+export * from './test-utils/seedTestWallet';
