+++ packages/wallet/src/transfer.test.ts
@@ -4,9 +4,8 @@ import { bn } from '@fuel-ts/math';
 import type { TransactionResultMessageOutReceipt } from '@fuel-ts/providers';
 import { Provider, ScriptTransactionRequest } from '@fuel-ts/providers';
 
-import { seedTestWallet, generateTestWallet } from '../test/utils';
-
 import { Wallet } from '.';
+import { seedTestWallet, generateTestWallet } from './test-utils';
 
 describe('Wallet', () => {
   it('can transfer a single type of coin to a single destination', async () => {
