--- packages/wallet/test/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './generateTestWallet';
-export * from './seedTestWallet';
