+++ packages/wallet/tsconfig.json
@@ -3,5 +3,5 @@
   "compilerOptions": {
     "outDir": "./dist"
   },
-  "include": ["src", "src/*.json", "test"]
+  "include": ["src"]
 }
